pyproject.toml
[project]
name = "langchain-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.0.2",
    "langchain-anthropic>=1.0.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "requests>=2.25.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "transformers>=4.21.0",
    "torch>=1.12.0",
    "huggingface-hub>=0.20.0",
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "langchain-openai>=1.0.1",
    "python-dotenv>=1.1.1",
]
